changed routes structure for each page
