--- src/routes/FAQ.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import NavBar from '/src/components/custom/NavBar.jsx'
-import FAQ1 from './FAQ/FAQ1'
-import Footer from '/src/components/custom/Footer.jsx'
-
-export default function ContactPage() {
-  return (
-    <>
-      <NavBar />
-      <FAQ1 />
-      <Footer />
-    </>
-  )
-}
