--- src/routes/Other/Notused/Offer.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-
-export default function Oferta() {
-  return (
-    <>
-      <div className="bg-background-primary flex lg:flex-row flex-col gap-6 items-center sm:gap-20 text-white lg:px-14 px-0 py-32 ">
-        <div className="flex lg:flex-row flex-col gap-6 mb-4 sm:mb-0">
-            <div className='flex flex-col justify-center items-center px-10'>
-                <h1 className="lg:text-6xl text-4xl font-bold items-start gap-6 ">Nie widzisz określonego produktu</h1>
-                <p className='text-xl mt-2 non-italic font-bold'>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse varius enim in eros elementum tristique.</p>
-            </div>
-        </div>
-  
-      </div>
-    </>
-  )
-}
