+++ src/routes/PrecisionFarming/_components/Cont3.jsx
