+++ src/routes/PrecisionFarming/_components/Cont4.jsx
