+++ src/routes/PrecisionFarming/_components/Cont5.jsx
