+++ src/routes/PrecisionFarming/_components/Cont6.jsx
