+++ src/routes/PrecisionFarming/_components/Map.jsx
