+++ src/routes/PrecisionFarming/_components/Map2.jsx
