+++ src/routes/PrecisionFarming/_components/Map3.jsx
