+++ src/routes/PrecisionFarming/_components/Map_time_slider.jsx
