+++ src/routes/PrecisionFarming/_components/SimpleMapComponent.jsx
