+++ src/routes/Products/page.jsx
@@ -1,6 +1,6 @@
 import NavBar from "src/components/custom/NavBar";
 import Footer from "@/layouts/Footer";
-import ContactUs from "./Contact/_components/ContactUs";
+import ContactUs from "../Contact/_components/ContactUs";
 import PriceTable from "src/components/custom/PriceTable";
