+++ src/routes/TermsOfUse/page.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import NavBar from '../custom/NavBar'
 import Footer from '../custom/Footer'
-import TermsOfUse1 from './TermsOfUse/TermsOfUse1'
+import TermsOfUse1 from './_components/TermsOfUse1'
 
 export default function TermsOfUse() {
   return (
